[python] make hb_profiler a top level module that only imports other modules
