+++ hammerblade/torch/tests/test_profiling.py
@@ -31,7 +31,7 @@ def test_execution_time_1():
     x = torch.randn(100000)
     y = x + x
     torch.hb_profiler.disable()
-    fancy = torch.hb_profiler.exec_time_fancy_print()
+    fancy = torch.hb_profiler.exec_time.fancy_print()
     assert fancy.find("aten::randn") != -1
     assert fancy.find("aten::add") != -1
     assert fancy.find("aten::ones") == -1
@@ -42,7 +42,7 @@ def test_execution_time_2():
     x = torch.randn(100000)
     y = x + x
     torch.hb_profiler.disable()
-    stack = torch.hb_profiler.exec_time_raw_stack()
+    stack = torch.hb_profiler.exec_time.raw_stack()
     assert stack.find("at::Tensor at::CPUType::{anonymous}::add(const at::Tensor&, const at::Tensor&, c10::Scalar)") != -1
     assert stack.find("at::Tensor at::TypeDefault::randn(c10::IntArrayRef, const c10::TensorOptions&)") != -1
     assert stack.find("at::Tensor at::TypeDefault::ones(c10::IntArrayRef, const c10::TensorOptions&)") == -1
@@ -55,7 +55,7 @@ def test_unimpl_1():
     x = torch.randn(100000)
     y = x + x
     torch.hb_profiler.disable()
-    unimpl = torch.hb_profiler.unimpl_print()
+    unimpl = torch.hb_profiler.unimpl.print()
     assert unimpl.find("aten::normal_") != -1
 
 def test_unimpl_2():
@@ -64,35 +64,35 @@ def test_unimpl_2():
     torch.hb_profiler.enable()
     y = x + x
     torch.hb_profiler.disable()
-    unimpl = torch.hb_profiler.unimpl_print()
+    unimpl = torch.hb_profiler.unimpl.print()
     assert unimpl.find("aten::normal_") == -1
 
 def test_chart_1():
     M = torch.randn(2, 3)
     mat1 = torch.randn(2, 3)
     mat2 = torch.randn(3, 3)
-    torch.hb_profiler.clear_beacon()
-    torch.hb_profiler.add_beacon("at::Tensor at::CPUType::{anonymous}::addmm(const at::Tensor&, const at::Tensor&, const at::Tensor&, c10::Scalar, c10::Scalar)")
+    torch.hb_profiler.chart.clear_beacon()
+    torch.hb_profiler.chart.add_beacon("at::Tensor at::CPUType::{anonymous}::addmm(const at::Tensor&, const at::Tensor&, const at::Tensor&, c10::Scalar, c10::Scalar)")
     torch.hb_profiler.enable()
     torch.add(M, mat1)
     torch.hb_profiler.disable()
-    torch.hb_profiler.clear_beacon()
-    chart = torch.hb_profiler.chart_print()
+    torch.hb_profiler.chart.clear_beacon()
+    chart = torch.hb_profiler.chart.print()
     assert chart == "[]\n"
 
 def test_chart_2():
     M = torch.randn(2, 3)
     mat1 = torch.randn(2, 3)
     mat2 = torch.randn(3, 3)
-    torch.hb_profiler.clear_beacon()
-    torch.hb_profiler.add_beacon("at::Tensor at::CPUType::{anonymous}::addmm(const at::Tensor&, const at::Tensor&, const at::Tensor&, c10::Scalar, c10::Scalar)")
+    torch.hb_profiler.chart.clear_beacon()
+    torch.hb_profiler.chart.add_beacon("at::Tensor at::CPUType::{anonymous}::addmm(const at::Tensor&, const at::Tensor&, const at::Tensor&, c10::Scalar, c10::Scalar)")
     torch.hb_profiler.enable()
     torch.add(M, mat1)
     torch.addmm(M, mat1, mat2)
     torch.addmm(M, mat1, mat2)
     torch.hb_profiler.disable()
-    torch.hb_profiler.clear_beacon()
-    chart = torch.hb_profiler.chart_print()
+    torch.hb_profiler.chart.clear_beacon()
+    chart = torch.hb_profiler.chart.print()
     golden = """[
     {
         "offload": false,
@@ -122,8 +122,8 @@ def test_route_1():
 ]
 """
     data = json.loads(route)
-    torch.hb_profiler.set_route_from_json(data)
-    _route = torch.hb_profiler.route_print()
+    torch.hb_profiler.route.set_route_from_json(data)
+    _route = torch.hb_profiler.route.print()
     assert _route == route
     torch.hb_profiler.enable()
     torch.addmm(M, mat1, mat2)
@@ -148,7 +148,7 @@ def test_route_2_F():
 ]
 """
     data = json.loads(route)
-    torch.hb_profiler.set_route_from_json(data)
+    torch.hb_profiler.route.set_route_from_json(data)
     torch.hb_profiler.enable()
     torch.addmm(M, mat1, mat2)
     torch.addmm(M, mat1, mat2)
